move the tour-overview and customize tour into containers
